+++ cli-tools/cli-lib/modules/constants-enumerate.js
@@ -1,3 +1,5 @@
+const path = require('path');
+
 const projectFileNames = {
   GENERATED_CONFIG: ['.aofl.json'],
   CONFIG: ['.aofl.js'],
@@ -9,7 +11,12 @@ const environments = {
   TEST: 'test'
 };
 
+const resources = {
+  CUSTOM_ELEMENTS_ES5_ADAPTER: path.join(__dirname, 'webpack-config', '__config', 'custom-elements-es5-adapter.js')
+}
+
 module.exports = {
   projectFileNames,
-  environments
+  environments,
+  resources
 };